fix: tests and config sync to use the signingkey variant of the authenticated delete, finish authenticated delete, add and update dependencies for compose and kotlin and add navigation library to make nested composables for create group and future easier to deal with. build preview avatar component
