+++ app/build.gradle
@@ -17,12 +17,12 @@ buildscript {
 plugins {
     id 'kotlin-kapt'
     id 'com.google.dagger.hilt.android'
+    id 'com.google.devtools.ksp' version "$kspVersion"
 }
 
 apply plugin: 'com.android.application'
 apply plugin: 'kotlin-android'
 apply plugin: 'witness'
-apply plugin: 'kotlin-kapt'
 apply plugin: 'kotlin-parcelize'
 apply plugin: 'kotlinx-serialization'
 apply plugin: 'dagger.hilt.android.plugin'
@@ -78,7 +78,7 @@ android {
         compose true
     }
     composeOptions {
-        kotlinCompilerExtensionVersion '1.4.7'
+        kotlinCompilerExtensionVersion '1.5.3'
     }
 
     defaultConfig {
@@ -206,7 +206,12 @@ android {
 dependencies {
 
     implementation("com.google.dagger:hilt-android:2.46.1")
-    kapt("com.google.dagger:hilt-android-compiler:2.44")
+    kapt("com.google.dagger:hilt-android-compiler:2.46")
+
+    implementation "io.github.raamcosta.compose-destinations:core:$composeDestinationsVersion"
+    ksp "io.github.raamcosta.compose-destinations:ksp:$composeDestinationsVersion"
+
+
 
     implementation "androidx.appcompat:appcompat:$appcompatVersion"
     implementation 'androidx.recyclerview:recyclerview:1.2.1'
@@ -330,22 +335,22 @@ dependencies {
     androidTestImplementation 'androidx.test.espresso:espresso-web:3.5.1'
     androidTestImplementation 'androidx.test.espresso.idling:idling-concurrent:3.5.1'
     androidTestImplementation 'androidx.test.espresso:espresso-idling-resource:3.5.1'
-    androidTestImplementation "androidx.compose.ui:ui-test-junit4:1.5.2"
-    debugImplementation "androidx.compose.ui:ui-test-manifest:1.5.2"
+    androidTestImplementation "androidx.compose.ui:ui-test-junit4:1.5.3"
+    debugImplementation "androidx.compose.ui:ui-test-manifest:1.5.3"
     androidTestUtil 'androidx.test:orchestrator:1.4.2'
 
     testImplementation 'org.robolectric:robolectric:4.4'
     testImplementation 'org.robolectric:shadows-multidex:4.4'
 
     implementation 'com.github.bumptech.glide:compose:1.0.0-alpha.5'
-    implementation 'androidx.compose.ui:ui:1.5.2'
-    implementation 'androidx.compose.ui:ui-tooling:1.5.2'
+    implementation 'androidx.compose.ui:ui:1.5.3'
+    implementation 'androidx.compose.ui:ui-tooling:1.5.3'
     implementation "com.google.accompanist:accompanist-themeadapter-appcompat:0.33.1-alpha"
     implementation "com.google.accompanist:accompanist-pager-indicators:0.33.1-alpha"
-    implementation "androidx.compose.runtime:runtime-livedata:1.5.2"
+    implementation "androidx.compose.runtime:runtime-livedata:1.5.3"
 
-    implementation 'androidx.compose.foundation:foundation-layout:1.5.2'
-    implementation 'androidx.compose.material:material:1.5.2'
+    implementation 'androidx.compose.foundation:foundation-layout:1.5.3'
+    implementation 'androidx.compose.material:material:1.5.3'
 }
 
 static def getLastCommitTimestamp() {
