+++ gradle.properties
@@ -18,7 +18,9 @@ org.gradle.jvmargs=-Xmx2048M -Dkotlin.daemon.jvm.options\="-Xmx2048M"
 org.gradle.unsafe.configuration-cache=true
 
 googleServicesVersion=4.3.12
-kotlinVersion=1.8.21
+kotlinVersion=1.9.10
+kspVersion=1.9.10-1.0.13
+composeDestinationsVersion=1.9.54
 android.useAndroidX=true
 appcompatVersion=1.6.1
 coreVersion=1.8.0
